requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.2
bert-extractive-summarizer==0.10.1
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
ffmpeg==1.4
filelock==3.13.1
fsspec==2023.12.1
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper @ git+https://github.com/openai/whisper.git@e58f28804528831904c3b6f2c0e473f346223433
packaging==23.2
preshed==3.0.9
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
pyTelegramBotAPI==4.14.0
pytube==15.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
setuptools-rust==1.8.1
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
srsly==2.4.8
sympy==1.12
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
